LayoutTests/http/tests/resourceLoadStatistics/website-data-removal-for-site-with-user-interaction-and-cross-origin-without-user-interaction.html

<!DOCTYPE html>
<html>
<head>
    <script src="/cookies/resources/cookie-utilities.js"></script>
    <script src="resources/util.js"></script>
</head>
<body onload="setTimeout('runTest()', 0)">
<div id="description">Check that script-writeable website data does not get removed for sites with user interaction, including partitioned cross-origin website data.</div>
<iframe src="http://localhost:8000/resourceLoadStatistics/resources/iframe-with-storage.html"></iframe>
<div id="output"></div>
<br>
<script>
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
    internals.settings.setStorageBlockingPolicy('BlockThirdParty');

    const iframe = document.getElementsByTagName("iframe")[0];
    let popup;

    const dbName = "TestDatabase";
    const objectStoreName = "TestObjectStore";
    const storageKey = "key";
    const storageValue = "value";

    const originUnderTest  = "http://127.0.0.1:8000";
    const crossOrigin = "http://localhost:8000";
    function runTest() {
        setEnableFeature(true, function () {
            testRunner.setStatisticsFirstPartyWebsiteDataRemovalMode(true, function() {
                testRunner.setStatisticsHasHadUserInteraction(originUnderTest, true, function() {
                    if (!testRunner.isStatisticsHasHadUserInteraction(originUnderTest))
                        addOutput("FAIL: " + originUnderTest + " did not get logged for user interaction.");
                    window.addEventListener("message", e => {
                        // Only the originUnderTest should have user interaction
                        testRunner.setStatisticsHasHadUserInteraction(crossOrigin, false, function() {
                            if (testRunner.isStatisticsHasHadUserInteraction(crossOrigin))
                                addOutput("FAIL: " + crossOrigin + " did not get logged for user interaction.");
                            else
                                addOutput("Opened cross-origin " + crossOrigin + " popup");
                            writeWebsiteDataAndContinue();
                        });
                    }, { once: true });
                    popup = window.open(crossOrigin + "/resourceLoadStatistics/resources/popup-with-storage.html");
                    if (popup == undefined)
                        addOutput("FAIL: Opening " + crossOrigin + " popup with opener was not successful.");
                    else
                        addOutput("Opening " + crossOrigin + " popup.");
                });
            });
        });
    }
</script>
</body>
</html>
